feat: add tests for breaking change detection in routes and type aliases
- Implemented unit tests for route detection in `patch-detection-routes.test.ts`, covering various scenarios including removed, changed, and added routes.
- Added tests for type alias detection in `patch-detection-type.test.ts`, focusing on union and record types.
- Created utility functions in `utils.ts` for generating test contexts and managing temporary folders.
- Introduced AST utility functions in `ast-utils.ts` to facilitate declaration finding and type reference handling.
- Updated changelog processing to reference the new breaking change detector structure.
- Enhanced TypeScript configuration to exclude specific directories from compilation.
