+++ eng/tools/js-sdk-release-tools-src/package.json
@@ -33,23 +33,33 @@
   "dependencies": {
     "@azure-tools/openapi-tools-common": "^1.2.2",
     "@npmcli/package-json": "^8.0.0",
+    "@typescript-eslint/scope-manager": "^7.17.0",
+    "@typescript-eslint/types": "^7.17.0",
+    "@typescript-eslint/typescript-estree": "^7.17.0",
+    "@typescript-eslint/utils": "^7.17.0",
+    "@typescript-eslint/visitor-keys": "^7.17.0",
     "colors": "1.4.0",
     "command-line-args": "^6.0.1",
+    "commander": "^12.1.0",
     "comment-json": "^4.2.5",
     "copyfiles": "^2.4.1",
     "fs-extra": "^11.3.0",
     "glob": "^13.0.6",
     "js-yaml": "^4.1.0",
+    "marked": "^13.0.2",
     "minimatch": "^10.2.4",
     "npm-registry-fetch": "^20.0.1",
     "parse-ts-to-ast": "^0.1.4",
+    "pino": "^9.3.1",
+    "pino-pretty": "^11.2.1",
     "semver": "^7.7.1",
     "shelljs": "^0.9.2",
     "simple-git": "^3.5.0",
     "string-template": "^1.0.0",
     "ts-morph": "^26.0.0",
     "tslib": "^2.8.1",
-    "typescript-codegen-breaking-change-detector": "0.7.10",
+    "typescript": "5.8.2",
+    "typescript-eslint": "^7.17.0",
     "unixify": "^1.0.0",
     "winston": "^3.17.0",
     "yaml": "^2.7.0"
@@ -66,7 +76,6 @@
     "prettier": "^3.8.3",
     "rimraf": "^6.0.1",
     "tsx": "^4.22.4",
-    "typescript": "5.8.2",
     "vitest": "4.1.9"
   },
   "peerDependencies": {
