--- eng/tools/js-sdk-release-tools-src/packages/typescript-codegen-breaking-change-detector/.npmrc
@@ -1,1 +1,0 @@
-registry=https://pkgs.dev.azure.com/azure-sdk/public/_packaging/azure-sdk-for-js/npm/registry/
