--- eng/tools/js-sdk-release-tools-src/packages/typescript-codegen-breaking-change-detector/README.md
@@ -1,11 +1,0 @@
-# typescript-codegen-breaking-change-detector
-
-> 🏗 Work in progress...
-
-Detect breaking changes to your TypeScript client generated by open api, typespec and more...
-
-## Contribution
-
-### sort-imports
-
-Install `amatiasq.sort-imports` VS code extension to sort imports automatically.
