--- eng/tools/js-sdk-release-tools-src/packages/typescript-codegen-breaking-change-detector/package.json
@@ -1,57 +1,0 @@
-{
-  "name": "typescript-codegen-breaking-change-detector",
-  "version": "0.7.10",
-  "description": "Detect breaking changes to your TypeScript client generated by open api, typespec and more...",
-  "main": "dist/src/index.js",
-  "types": "./dist/src/index.d.ts",
-  "files": [
-    "dist"
-  ],
-  "scripts": {
-    "build": "rimraf ./dist && tsc",
-    "build:watch": "rimraf ./dist && tsc --watch",
-    "start:dev": "tsx watch src/cli.ts",
-    "start": "npm run build && node dist/index.js",
-    "test": "vitest --run",
-    "test:watch": "vitest",
-    "prettier-format": "run-script-os",
-    "prettier-format:win32": "prettier --config .prettierrc \"./src/**/*.ts\" --write",
-    "prettier-format:darwin:linux": "prettier --config .prettierrc 'src/**/*.ts' --write",
-    "prettier-format:default": "prettier --config .prettierrc 'src/**/*.ts' --write",
-    "prettier-watch": "run-script-os",
-    "prettier-watch:win32": "onchange \"src/**/*.ts\" -- prettier --write {{changed}}",
-    "prettier-watch:darwin:linux": "onchange 'src/**/*.ts' -- prettier --write {{changed}}",
-    "prettier-watch:default": "onchange 'src/**/*.ts' -- prettier --write {{changed}}"
-  },
-  "license": "MIT",
-  "dependencies": {
-    "@typescript-eslint/scope-manager": "^7.17.0",
-    "@typescript-eslint/types": "^7.17.0",
-    "@typescript-eslint/typescript-estree": "^7.17.0",
-    "@typescript-eslint/utils": "^7.17.0",
-    "@typescript-eslint/visitor-keys": "^7.17.0",
-    "commander": "^12.1.0",
-    "fs-extra": "^11.2.0",
-    "glob": "^11.0.0",
-    "marked": "^13.0.2",
-    "pino": "^9.3.1",
-    "pino-pretty": "^11.2.1",
-    "ts-morph": "^26.0.0",
-    "typescript": "5.8.2",
-    "typescript-eslint": "^7.17.0"
-  },
-  "devDependencies": {
-    "@types/fs-extra": "^11.0.4",
-    "husky": "^9.0.11",
-    "onchange": "^7.1.0",
-    "prettier": "^3.8.0",
-    "rimraf": "^5.0.7",
-    "tsx": "^4.16.2",
-    "vitest": "^4.1.0"
-  },
-  "husky": {
-    "hooks": {
-      "pre-commit": "npm run prettier-format && npm run lint"
-    }
-  }
-}
