--- eng/tools/js-sdk-release-tools-src/packages/typescript-codegen-breaking-change-detector/src/common/models/rules/rule-descriptions.ts
@@ -1,5 +1,0 @@
-import { ignoreInlineDeclarationsInOperationGroup } from "./rule-ids";
-
-export const ruleDescriptions: { [ruleId: string]: string } = {
-  [ignoreInlineDeclarationsInOperationGroup]: "Ignore inline types in routes",
-};
