--- eng/tools/js-sdk-release-tools-src/packages/typescript-codegen-breaking-change-detector/src/index.ts
@@ -1,4 +1,0 @@
-export * from "./azure/detect-breaking-changes";
-export * from "./azure/common/types";
-export * from "./azure/patch/patch-detection";
-export * from "./utils/ast-utils";
