--- eng/tools/js-sdk-release-tools-src/packages/typescript-codegen-breaking-change-detector/vitest.config.ts
@@ -1,8 +1,0 @@
-import { defineConfig } from "vitest/config";
-
-export default defineConfig({
-  test: {
-    testTimeout: 30000,
-    exclude: ["misc/**", "dist/**", "node_modules/**"],
-  },
-});
