+++ eng/tools/js-sdk-release-tools-src/src/breaking-change-detector/common/models/rules/rule-descriptions.ts
@@ -1,0 +1,5 @@
+import { ignoreInlineDeclarationsInOperationGroup } from "./rule-ids.js";
+
+export const ruleDescriptions: { [ruleId: string]: string } = {
+  [ignoreInlineDeclarationsInOperationGroup]: "Ignore inline types in routes",
+};
