+++ eng/tools/js-sdk-release-tools-src/src/breaking-change-detector/index.ts
@@ -1,0 +1,4 @@
+export * from "./azure/detect-breaking-changes.js";
+export * from "./azure/common/types.js";
+export * from "./azure/patch/patch-detection.js";
+export * from "./utils/ast-utils.js";
