+++ eng/tools/js-sdk-release-tools-src/src/breaking-change-detector/misc/test-cases/rest-level-client-to-rest-level-client/current-package/tsconfig.json
